support more num literals
similar to the ones in clojure (and the regex is very much inspired by
the one in tools.reader), e.g. hexadecimal, octal and my personal
favourite: radix are all there.

i think this breaks when the parsing fails, but it parses different
numbers for now and addition still works.