add a makefile
i want this because this allows me to specify which python to use,
because my system doesn't have a python2 as it's default python.
otherwise i'd have to change all the scripts.

it also has the advantage of having all the utilities in one place and
even a "help" command.