represent ^:examples as a vector of vectors
an example is a vector of [expr output val], where expr is a string
containing a read-string'able expression, output is a string containing
the output that is printed when evaluating the expr and val is the value
returned.

this allows checking the examples automatically, prettier doc output and
potential use of this metadata.