fix math ffi on linux
On Linux, libm (and other libraries) are separate, so we have to use
different libraries. One of these is libm, which contains things like
`floor`, `atan2` and quite a few other things.

There should be symlinks from `libm` et al to `libSystem` on Mac [1], so
it should still work there.

[1]: http://www.finkproject.org/doc/porting/porting.en.html#basics.libraries