Fix line comment bug and improve reader tests
Previously, code like
(1 2 3 ;foo
)
would fail to parse, because LineCommentReader would attempt to return
the next form, which was `)`, causing an error.

To fix this I changed the semantics of `read` to allow it to not return
a form at all (signalled by returning the reader itself, as in
https://github.com/clojure/tools.reader), for situations like `;`.
Unsure if this is the best approach to fixing, open to discussion.
Another option could be to treat comments as whitespace.

Also, the reader tests had some cases where they wouldn't verify the
input properly, so I fixed those as well.