Implement map without using iterators, fixes #242
The fact that map was implemented with iterators caused the following to
happen:

```
user => (def sqr #(* % %))
<inst pixie.stdlib.Var>
user => (def sqrs (map sqr [1 2 3]))
<inst pixie.stdlib.Var>
user => sqrs
<inst pixie.stdlib.ShallowContinuation>
user => (vec sqrs)
[1 4 9]
user => (vec sqrs)
[]
```

This commit implements map differently and adds tests to ensure
that it behaves as intended.

The implementation is pretty similar to the CLJS one, there are just a
few things that aren't available (since `map` is defined earlier in the
pixie stdlib than the CLJS core).