Improve assert-throws? macro
Since `assert-throws?` throws an exception inside the try regardless of
if the body throws, if the body *doesn't* throw, you get a sort of
strange error message:

```
RuntimeException: Assert failed Expected message: exception! but got
Assert failed Expected a <type pixie.stdlib.RuntimeException>
exception: exception!
```

I fixed this, improved the messages a bit, and also added some extra
arities for just checking that an expression throws, and also just
checking the class of the exception, instead of always the class and the
message.