Implement -seq for sets the same way as maps
Sets were using iterators instead of just building a list (which were
mentioned should be removed in #242), and I guess some of the iterator
code doesn't work properly for maps, because executing `(seq #{4 7})`
causes a segfault without this patch.

I'm not really familar with the implementation of iterators over maps,
but this is sort of an interesting bug because I've only been able to
reproduce it with sets that have #{4 7} as a subset.