Add `eduction` and `completing`
The `eduction` function was added, as requested in issue #483.

The implementation and the tests were imported and adapted from
Clojure's standard library (from [1] and [2], respectively).

Since Clojure's implementation uses it, the function `completing` was
imported as well (from the same source).

As of now, Pixie's stdlib lacks a `sort`/`sort-by` function, one of the
tests is disabled (commented out), and another is modified in order to
avoid the call to `sort-by`.

Sources:
Both from Clojure's main repo, fetched at commit `3b98a00`.
[1] https://github.com/clojure/clojure/blob/3b98a00e86f961550fb2eaee64e70754e04d1089/src/clj/clojure/core.clj#L7339
[2] https://github.com/clojure/clojure/blob/3b98a00e86f961550fb2eaee64e70754e04d1089/test/clojure/test_clojure/transducers.clj